requirements.txt
aiohttp==3.10.5
anthropic==0.34.0
fastapi==0.112.1
grpcio==1.65.5
grpcio-tools==1.65.5
grpcio-reflection==1.65.5
pydantic==2.8.2
python-dotenv==1.0.1
uvicorn==0.30.6
pyyaml==6.0.2
openai==1.41.0
scalar-fastapi==1.0.1
